requirements.txt
fastapi
uvicorn
supabase
openai
anthropic
pypdf2
python-dotenv
python-multipart
PyJWT
